// FacetsUI/src/app/modules/team-member/models/team-member-search.model.ts
export class TeamMemberSearchModel {
    isRegisteredToFacets: boolean;
    firstName: string;
    lastName: string;
    nicNumber: string;
    passportNumber: string;
    phoneNumber: string;
    email: string;
    teamMemberId: string;
    countryId: string;
    passCategoryId: string;
    imageURL: string;
    status: string;
    identityType: string;
}